services/auth-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
bcrypt
python-jose[cryptography]
python-dotenv
pydantic[email]